Add systemd user unit: pat.service
This is started by the default.target, which is similar to the
multi-user.target of the system service manager.

There is no dependency on the system ax25.service unit, because
I could not figure out a way to do that yet. Therefore there are
probably race conditions on boot if ax25 is used. Not ideal.